librmnetctl: Proper error handling of if_nametoindex()
if_nametoindex() returns an unsigned integer, so errors are indicated
by returning 0 to the caller. This updates the error handling to
check for 0 instead of negative numbers to avoid passing bogus
netlink messages to the kernel. This allows for more consistent and
helpful error messages to be returned to the caller

Change-Id: I5f78781c973e7cb5a8bbc813a31cd49e433034a4